Convert baseColorTextureIndices to uin32_t for GPU consumption

+++ src/common/gltf_model.hpp
@@ -5,6 +5,7 @@
 
 #include <glm/glm.hpp>
 
+#include <cstdint>
 #include <filesystem>
 #include <span>
 #include <vector>
@@ -16,20 +17,20 @@ class GltfModel
 public:
     GltfModel(std::filesystem::path gltfPath);
 
-    std::span<const Positions>   positions() const { return mPositions; }
-    std::span<const Normals>     normals() const { return mNormals; }
-    std::span<const TexCoords>   texCoords() const { return mTexCoords; }
-    std::span<const std::size_t> baseColorTextureIndices() const
+    std::span<const Positions>     positions() const { return mPositions; }
+    std::span<const Normals>       normals() const { return mNormals; }
+    std::span<const TexCoords>     texCoords() const { return mTexCoords; }
+    std::span<const std::uint32_t> baseColorTextureIndices() const
     {
         return mBaseColorTextureIndices;
     }
     std::span<const Texture> baseColorTextures() const { return mBaseColorTextures; }
 
 private:
-    std::vector<Positions>   mPositions;
-    std::vector<Normals>     mNormals;
-    std::vector<TexCoords>   mTexCoords;
-    std::vector<std::size_t> mBaseColorTextureIndices;
+    std::vector<Positions>     mPositions;
+    std::vector<Normals>       mNormals;
+    std::vector<TexCoords>     mTexCoords;
+    std::vector<std::uint32_t> mBaseColorTextureIndices;
 
     std::vector<Texture> mBaseColorTextures;
 };
